# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Media Downloader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: download
    description: Media download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /download/info:
    post:
      operationId: getMediaInfo
      tags: [download]
      summary: Get available formats/qualities for a media URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInfoInput"
      responses:
        "200":
          description: Media info with available formats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaInfo"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download/start:
    post:
      operationId: startDownload
      tags: [download]
      summary: Start a download job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadInput"
      responses:
        "202":
          description: Download job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadJob"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download/{jobId}/status:
    get:
      operationId: getDownloadStatus
      tags: [download]
      summary: Get download job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download/{jobId}/file:
    get:
      operationId: downloadFile
      tags: [download]
      summary: Download the completed ZIP file
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ZIP file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: Job not found or not complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MediaInfoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: URL of the media to download

    MediaFormat:
      type: object
      required: [formatId, quality, ext, filesize]
      properties:
        formatId:
          type: string
        quality:
          type: string
        ext:
          type: string
        filesize:
          type: ["integer", "null"]
        vcodec:
          type: ["string", "null"]
        acodec:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]

    MediaInfo:
      type: object
      required: [title, thumbnail, platform, duration, formats]
      properties:
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
        platform:
          type: string
        duration:
          type: ["integer", "null"]
        formats:
          type: array
          items:
            $ref: "#/components/schemas/MediaFormat"

    DownloadInput:
      type: object
      required: [url, formatId]
      properties:
        url:
          type: string
        formatId:
          type: string
        title:
          type: ["string", "null"]

    DownloadJob:
      type: object
      required: [jobId, status, createdAt]
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, downloading, zipping, done, error]
        progress:
          type: ["integer", "null"]
        error:
          type: ["string", "null"]
        filename:
          type: ["string", "null"]
        createdAt:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string