# lib/api-zod/src/generated/types/downloadJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Media Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadJobStatus } from './downloadJobStatus';

export interface DownloadJob {
  jobId: string;
  status: DownloadJobStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  filename?: string | null;
  createdAt: string;
}