# lib/api-zod/src/generated/types/mediaInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Media Downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFormat } from './mediaFormat';

export interface MediaInfo {
  title: string;
  /** @nullable */
  thumbnail: string | null;
  platform: string;
  /** @nullable */
  duration: number | null;
  formats: MediaFormat[];
}