# lib/api-zod/src/generated/types/mediaInfoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Media Downloader API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaInfoInput {
  /** URL of the media to download */
  url: string;
}